partially added tandem model
discovered i need to add a navigation model
