+++ src/classes/mesh/tandem.py
@@ -1,15 +1,15 @@
-from OCC.Core.TopoDS import TopoDS_Shape, TopoDS_Compound
-from OCC.Core.BRepPrimAPI import BRepPrimAPI_MakeCylinder, BRepPrimAPI_MakeSphere, BRepPrimAPI_MakePrism
-from OCC.Core.BRepBuilderAPI import BRepBuilderAPI_MakeEdge, BRepBuilderAPI_MakeWire, BRepBuilderAPI_MakeFace
-from OCC.Core.BRepOffsetAPI import BRepOffsetAPI_ThruSections
-from OCC.Core.BRepFill import BRepFill_PipeShell
-from OCC.Core.BRepAlgoAPI import BRepAlgoAPI_Fuse
+import math
+import numpy as np
+
 from OCC.Core.BRep import BRep_Builder
+from OCC.Core.BRepAlgoAPI import BRepAlgoAPI_Fuse
+from OCC.Core.BRepBuilderAPI import BRepBuilderAPI_MakeEdge, BRepBuilderAPI_MakeWire, BRepBuilderAPI_MakeFace
+from OCC.Core.BRepFill import BRepFill_PipeShell
+from OCC.Core.BRepOffsetAPI import BRepOffsetAPI_ThruSections
+from OCC.Core.BRepPrimAPI import BRepPrimAPI_MakeCylinder, BRepPrimAPI_MakeSphere, BRepPrimAPI_MakePrism
 from OCC.Core.GC import GC_MakeArcOfCircle
 from OCC.Core.gp import *
-
-import numpy as np
-import math
+from OCC.Core.TopoDS import TopoDS_Shape, TopoDS_Compound
 
 
 # generates a custom tandem with inputed values
@@ -17,10 +17,9 @@ def generate_tandem(
         channel_diameter: float = 4.0,
         tip_diameter: float = 8.0,
         tip_thickness: float = 4.0,
-        tip_angle = 60.0,
+        tip_angle=60.0,
         tip_height: float = 160.0) -> TopoDS_Shape:
-
-    """Calculate the edges needs on a XZ plane where we pretend Z is Y"""    
+    """Calculate the edges needs on a XZ plane where we pretend Z is Y"""
     # variables used
     tip_rads = math.radians(tip_angle)
     curve_offset = 15.0  # used for the curve
@@ -29,18 +28,19 @@ def generate_tandem(
 
     # points
     p0 = [0, 0, 0]  # origin
-    p1 = [0, 0, tip_height] # where the tip begins
+    p1 = [0, 0, tip_height]  # where the tip begins
 
     # creating a curved edge, so the next point needs to be [cos(angle), 0, sin(angle)]
-    x = curve_offset - (math.cos(tip_rads) * curve_offset) 
+    x = curve_offset - (math.cos(tip_rads) * curve_offset)
     z = math.sin(tip_rads) * curve_offset  # curve offset sets the length
     p2 = [x, 0, z + tip_height]
 
     # tip continues along where the curve ends
-    angle_rads = math.radians(90 - tip_angle)  # the angle at the end of the curve
+    # the angle at the end of the curve
+    angle_rads = math.radians(90 - tip_angle)
     x = math.cos(angle_rads)
     z = math.sin(angle_rads)
-    vec = gp_Vec(x, 0 , z) * tip_thickness
+    vec = gp_Vec(x, 0, z) * tip_thickness
     p3 = [p2[0] + vec.X(), 0, p2[2] + vec.Z()]
 
     # lines to be used later
@@ -51,28 +51,32 @@ def generate_tandem(
     # tip's base
     p4 = line2.get_point_distance_from(p2, -tip_radius)  # pos x end
     p6 = line2.get_point_from_x(-channel_radius)  # neg x end
-    p5 = line2.get_point_distance_from(p6, -tip_radius) 
+    p5 = line2.get_point_distance_from(p6, -tip_radius)
 
     edges = []
     edges.append(make_arc_from_points(p2, p4, tip_radius))
-    edges.append(make_edge([p2[0], -tip_radius, p2[2]], [p5[0], -tip_radius, p5[2]]))
+    edges.append(make_edge([p2[0], -tip_radius, p2[2]],
+                 [p5[0], -tip_radius, p5[2]]))
     edges.append(make_arc_from_points(p5, p6, tip_radius))
-    edges.append(make_edge([p2[0], tip_radius, p2[2]], [p5[0], tip_radius, p5[2]]))
+    edges.append(make_edge([p2[0], tip_radius, p2[2]],
+                 [p5[0], tip_radius, p5[2]]))
 
     wires = []
     wires.append(make_wire(edges))
 
-    #tip's end
+    # tip's end
     p7 = line3.get_point_distance_from(p3, -tip_radius)  # pos x boundry
     p9 = line3.get_point_from_x(-channel_radius)  # neg x boundry
     p8 = line3.get_point_distance_from(p9, -tip_radius)
 
     edges = []
     edges.append(make_arc_from_points(p3, p7, tip_radius))
-    edges.append(make_edge([p3[0], -tip_radius, p3[2]], [p8[0], -tip_radius, p8[2]]))
+    edges.append(make_edge([p3[0], -tip_radius, p3[2]],
+                 [p8[0], -tip_radius, p8[2]]))
     edges.append(make_arc_from_points(p8, p9, tip_radius))
-    edges.append(make_edge([p3[0], tip_radius, p3[2]], [p8[0], tip_radius, p8[2]]))
-    wires.append(make_wire(edges))  
+    edges.append(make_edge([p3[0], tip_radius, p3[2]],
+                 [p8[0], tip_radius, p8[2]]))
+    wires.append(make_wire(edges))
 
     # base filling in
     edges = []
@@ -87,17 +91,9 @@ def generate_tandem(
     shapes = []
     shapes.append(make_cylinder(p0, p6, channel_radius))
     shapes.append(make_symmetrical_shape(wire, channel_radius))
-    shapes.append(make_curved_pipe(p1, make_arc(p1, p2), [0, 0, 1], channel_radius))
+    shapes.append(make_curved_pipe(
+        p1, make_arc(p1, p2), [0, 0, 1], channel_radius))
     shapes.append(make_thru_shape(wires))
-
-    #result = TopoDS_Compound()
-    #builder = BRep_Builder()
-    #builder.MakeCompound(result)
-    #for wire in wires:
-    #    builder.Add(result, wire)
-    #for shape in shapes:
-    #    builder.Add(result, shape)
-    #    pass
 
     return fuse_shapes(shapes)
 
@@ -111,7 +107,7 @@ def make_edge(p1, p2):
 def make_arc(start_point, end_point):
     p1 = gp_Pnt(start_point[0], start_point[1], start_point[2])
     p2 = gp_Pnt(end_point[0], end_point[1], end_point[2])
-    direction = gp_Vec(0,0,1)
+    direction = gp_Vec(0, 0, 1)
     arc = GC_MakeArcOfCircle(p1, direction, p2).Value()
     return BRepBuilderAPI_MakeEdge(arc).Edge()
 
@@ -129,7 +125,7 @@ def make_point(point):
     return BRepPrimAPI_MakeSphere(p, 1.0).Shape()
 
 
-def make_wire(edges:[]):
+def make_wire(edges: []):
     wire = BRepBuilderAPI_MakeWire()
     for edge in edges:
         wire.Add(edge)
@@ -159,7 +155,8 @@ def make_symmetrical_shape(wire, distance):
 def make_curved_pipe(start_point, arc, start_direction, radius):
     p1 = gp_Pnt(start_point[0], start_point[1], start_point[2])
     wire = BRepBuilderAPI_MakeWire(arc).Wire()
-    direction = gp_Dir(start_direction[0], start_direction[1], start_direction[2])
+    direction = gp_Dir(start_direction[0],
+                       start_direction[1], start_direction[2])
     axis = gp_Ax2(p1, direction)
     circle = gp_Circ(axis, radius)
     circle_edge = BRepBuilderAPI_MakeEdge(circle).Edge()
@@ -178,14 +175,14 @@ def fuse_shapes(shapes: []):
         if result is None:
             result = shape
             continue
-        
+
         result = BRepAlgoAPI_Fuse(result, shape).Shape()
-    
+
     return result
 
 
 class Line:
-    def __init__(self, *, p1 = None, p2 = None, m: float = None, b: float = None):
+    def __init__(self, *, p1=None, p2=None, m: float = None, b: float = None):
         self.slope = m or Line._get_slope(p1, p2)
         self.y_intercept = b or Line._get_y_intercept(self.slope, p1)
 
@@ -193,32 +190,35 @@ class Line:
         """m = y2 - y1 / x2 - x1"""
         rise = p2[2] - p1[2]
         run = p2[0] - p1[0]
-        return rise / run 
-    
+        return rise / run
+
     def _get_y_intercept(slope: float, p1: []):
         return p1[2] - (slope * p1[0])
-    
-    def get_perpindicular_line(self, p1:[]):
+
+    def get_perpindicular_line(self, p1: []):
         slope = self.slope
         m = -1 / slope
         b = Line._get_y_intercept(m, p1)
         return Line(m=m, b=b)
 
-    def get_point_from_x(self, x:float) -> []:
+    def get_point_from_x(self, x: float) -> []:
         m = self.slope
         b = self.y_intercept
         return [x, 0, m * x + b]
-    
-    def get_point_from_y(self, y:float) -> []:
+
+    def get_point_from_y(self, y: float) -> []:
         m = self.slope
         b = self.y_intercept
         return [(y - b) / m, 0, y]
-    
+
     def get_point_distance_from(self, p1: [], distance: float) -> []:
         value = distance / math.sqrt(1 + self.slope**2)
-        if distance > 0: x = p1[0] + value
-        else: x = p1[0] - value
+        if distance > 0:
+            x = p1[0] + value
+        else:
+            x = p1[0] - value
         return self.get_point_from_x(x)
-    
+
+
 if __name__ == "__main__":
     generate_tandem()
